ats-resume-checker/requirements.txt
flask
google-genai
python-dotenv
PyPDF2
python-docx